requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
Cython==3.0.6
docopt==0.6.2
idna==3.6
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pipreqs==0.4.13
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
six==1.16.0
tesseract==0.1.3
tesseract-ocr==0.0.1
tzdata==2023.3
urllib3==2.1.0
yarg==0.1.9
